backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.0
yfinance>=0.2.35
pandas>=2.2.0
numpy>=1.26.0,<2.0.0
scikit-learn>=1.5.0
tensorflow>=2.15.0
python-multipart>=0.0.6
httpx>=0.27.0
